♻️ refactor: chat 도메인 관련 파일이 늘어남에 따라 일관성을 위해 디렉토리 추가
