+++ server/src/chat/module/chat.module.ts
@@ -1,8 +1,8 @@
 import { Module } from '@nestjs/common';
-import { ChatGateway } from './chat.gateway';
+import { ChatGateway } from '../chat.gateway';
 import { ScheduleModule } from '@nestjs/schedule';
-import { ChatService } from './chat.service';
-import { ChatScheduler } from './chat.scheduler';
+import { ChatService } from '../service/chat.service';
+import { ChatScheduler } from '../scheduler/chat.scheduler';
 
 @Module({
   imports: [ScheduleModule.forRoot()],
